Properly extend Error instead of implementing it, update the Error message and stack to appear in the console and error activity
